+++ src/pages/login/components/form-options.tsx
@@ -1,4 +1,4 @@
-import paths from '@/configs/paths'
+import paths from '@/routes/paths'
 import { Button, Stack, Typography } from '@mui/material'
 import { Link } from 'react-router-dom'
 
@@ -8,7 +8,7 @@ export function FormOptions() {
       <Button component={Link} to={paths.root} variant="plain" disableRipple>
         <Typography variant="caption">아이디 / 비밀번호 찾기</Typography>
       </Button>
-      <Button component={Link} to={paths.signup} variant="plain" disableRipple>
+      <Button component={Link} to={paths.join} variant="plain" disableRipple>
         <Typography variant="caption">회원가입</Typography>
       </Button>
     </Stack>
