+++ src/pages/login/index.ts
@@ -1,0 +1,3 @@
+import { LoginPage } from './login-page'
+
+export default LoginPage
