--- src/pages/sign-in/index.ts
@@ -1,3 +1,0 @@
-import { SignInPage } from './sign-in-page'
-
-export default SignInPage
