+++ src/routes/router.tsx
@@ -1,7 +1,7 @@
-import paths from '@/configs/paths'
+import paths from '@/routes/paths'
 import MainLayout from '@/layouts/main-layout'
 import MainPage from '@/pages/main'
-import SignInPage from '@/pages/sign-in'
+import LoginPage from '@/pages/login'
 import { createBrowserRouter } from 'react-router-dom'
 
 const router = createBrowserRouter([
@@ -13,8 +13,8 @@ const router = createBrowserRouter([
         element: <MainPage />,
       },
       {
-        path: paths.signin,
-        element: <SignInPage />,
+        path: paths.login,
+        element: <LoginPage />,
       },
     ],
   },
